improve code coverage of mailman
