+++ tests/data/listserv.3GPP.txt
@@ -1,0 +1,1 @@
+https://list.etsi.org/scripts/wa.exe?A0=3GPP_TSG_RAN_WG2_MBMS_RRC
