+++ tests/data/listserv.IEEE.txt
@@ -1,0 +1,10 @@
+https://listserv.ieee.org/cgi-bin/wa?A0=STDS-802-PRIVACY
+https://listserv.ieee.org/cgi-bin/wa?A0=STDS-802-1-MINUTES
+https://listserv.ieee.org/cgi-bin/wa?A0=STDS-802-15-14
+https://listserv.ieee.org/cgi-bin/wa?A0=TORONTO-YP
+https://listserv.ieee.org/cgi-bin/wa?A0=STDS-802-15-15
+https://listserv.ieee.org/cgi-bin/wa?A0=IEEE-TEST
+https://listserv.ieee.org/cgi-bin/wa?A0=IEEE-OC-Jobs
+https://listserv.ieee.org/cgi-bin/wa?A0=GamesSIG
+https://listserv.ieee.org/cgi-bin/wa?A0=EREFLECTOR
+https://listserv.ieee.org/cgi-bin/wa?A0=IEEEOREGON-ANNOUNCE
